Move DirectoryFilters to ui module
